requirements.txt
discord.py>=2.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.0 